pyproject.toml
[tool.poetry]
name = "sanic-pastebin"
version = "0.1.0"
description = ""
authors = ["darialissi"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
sanic = {extras = ["ext"], version = "^24.6.0"}
SQLAlchemy = "^2.0.34"
pydantic-settings = "^2.5.2"
asyncpg = "^0.29.0"
bcrypt = "^4.2.0"
pyjwt = {extras = ["crypto"], version = "^2.9.0"}
jinja2 = "^3.1.4"
aiobotocore = "^2.15.0"
redis = "^5.0.8"
uvicorn = {extras = ["standard"], version = "^0.30.6"}

[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
isort = "^5.13.2"
flake8-pyproject = "^1.2.3"

[tool.black]
line-length = 120
target_version = ["py311"]

[tool.isort]
profile = "black"
py_version = 311

[tool.flake8]
max-line-length = 120

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
